divide en sidebar en components
