+++ dashWhat/src/utils/orderStatus.js
@@ -1,9 +1,11 @@
-export const ORDER_STATUS_COLORS = {
-  'No leido': 'bg-gray-500 text-white',
-  'Sin Confirmar': 'bg-blue-500 text-white',
-  'Confirmado': 'bg-purple-500 text-white',
+export const ORDER_STATUS = {
+  'No leido': 'bg-red-500 text-white',
+  'Sin Confirmar': 'bg-red-500 text-white',
+  'Confirmado': 'bg-green-500 text-white',
   'En proceso': 'bg-yellow-500 text-black',
-  'Entregado': 'bg-green-500 text-white',
-  'Cancelado': 'bg-red-500 text-white',
+  'Entregado': 'bg-blue-500 text-white',
+  'Cancelado': 'bg-gray-800 text-white',
   // Añade más estados y colores según sea necesario
 };
+
+
